Multithread final presentation

+++ tableau.h
@@ -3,10 +3,10 @@
 #ifndef TABLEAU_H
 #define TABLEAU_H
 
-void rowcopy(QState *q, long i, long k);
-void rowswap(QState *q, long i, long k);
-void rowset(QState *q, long i, long b);
-void rowmult(QState *q, long i, long k);
-int clifford(QState *q, long i, long k);
+void rowcopy(QState *q, int i, int k);
+void rowswap(QState *q, int i, int k);
+void rowset(QState *q, int i, int b);
+void rowmult(QState *q, int i, int k);
+int clifford(QState *q, int i, int k);
 
 #endif
